refactor: setup stores
