+++ frontend/src/hooks/index.ts
@@ -9,3 +9,4 @@ export * from './useInputMap';
 export * from './useContractQueries';
 export * from './usePythonSyntax';
 export * from './useFileName';
+export * from './useSetupStores';
